[THE-1] Added countdown to site (#1)
Per feature requests on the #ot-theresa-is-in-town Slack channel, this commit adds a countdown feature to the site.

This commit already includes usage of the countdown. To use the countdown for following interations of the event, 1. include countdown.js on the site, and 2. pass two parameters, a date string in ISO format and the HTML element where the countdown needs to be displayed.